feat: DMList에 스켈레톤 추가
