+++ src/view/layout/sidebar/components/skeletons/SkeletonDMList.tsx
@@ -1,0 +1,15 @@
+const SkeletonDMList = () => {
+  return (
+    <div className='flex flex-col items-start gap-[2px] w-[203px]'>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+      <div className='border-[#3E4147] bg-[#3E4147] h-[42px] rounded-[8px] w-full animate-pulse'></div>
+    </div>
+  );
+};
+
+export default SkeletonDMList;
